requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
arrow==1.3.0
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via -r requirements.in
black==23.12.1
    # via -r requirements.in
blurhash==1.1.4
    # via mastodon-py
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via mastodon-py
defusedxml==0.7.1
    # via untangle
execnet==2.0.2
    # via pytest-xdist
faker==22.5.1
    # via pytest-faker
feedparser==6.0.11
    # via -r requirements.in
future==0.18.3
    # via python-twitter
idna==3.6
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
markdown-it-py==3.0.0
    # via rich
mastodon-py==1.8.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pyright
numpy==1.26.3
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via
    #   black
    #   build
    #   pytest
pandas==2.2.0
    # via -r requirements.in
pathspec==0.12.1
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via black
pluggy==1.4.0
    # via pytest
pygments==2.17.2
    # via rich
pyproject-hooks==1.0.0
    # via build
pyright==1.1.348
    # via -r requirements.in
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-xdist
pytest-faker==2.0.0
    # via -r requirements.in
pytest-xdist==3.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   faker
    #   mastodon-py
    #   pandas
python-frontmatter==1.1.0
    # via -r requirements.in
python-magic==0.4.27
    # via mastodon-py
python-slugify==8.0.1
    # via -r requirements.in
python-twitter==3.5
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   python-frontmatter
    #   yamale
requests==2.31.0
    # via
    #   -r requirements.in
    #   mastodon-py
    #   python-twitter
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via python-twitter
rich==13.7.0
    # via -r requirements.in
ruamel-yaml==0.18.5
    # via -r requirements.in
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   mastodon-py
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
text-unidecode==1.3
    # via python-slugify
types-python-dateutil==2.8.19.20240106
    # via arrow
tzdata==2023.4
    # via pandas
untangle==1.2.1
    # via -r requirements.in
urllib3==2.1.0
    # via requests
wheel==0.42.0
    # via pip-tools
yamale==4.0.4
    # via -r requirements.in
yarl==1.9.4
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
